+++ lib/models/config_model.dart
@@ -1,3 +1,5 @@
+import 'package:args/args.dart';
+
 /// Config model
 class ConfigModel {
   /// percentage
@@ -11,9 +13,20 @@ class ConfigModel {
   /// package name
   final String? packageName;
 
+  /// Command
+  final ArgResults? command;
+
+  // If should include untested files to report
+  final bool includeUntestedFiles;
+
+  final String? lcovGen;
+
   ConfigModel(
       {required this.percentage,
       required this.excludeSuffixes,
+      required this.includeUntestedFiles,
       required this.log,
-      this.packageName});
+      this.packageName,
+      this.command,
+      this.lcovGen});
 }